+++ packages/core/mashroom/src/config/MashroomServerConfigLoader.js
@@ -4,9 +4,9 @@ import os from 'os';
 import path from 'path';
 import {existsSync} from 'fs';
 import {createReadonlyProxy} from '@mashroom/mashroom-utils/lib/readonly_utils';
+import {evaluateTemplatesInConfigObject} from '@mashroom/mashroom-utils/lib/config_utils';
 import {deepAssign} from '@mashroom/mashroom-utils/lib/model_utils';
 import defaultConfig from './mashroom_default_config';
-import {evaluateTemplatesInConfigObject} from '../utils/config_utils';
 import ServerConfigurationError from '../errors/ServerConfigurationError';
 
 import type {MashroomLogger, MashroomLoggerFactory} from '../../type-definitions';