+++ packages/core/mashroom/src/server/routes/mashroom/admin/escape_html.js
@@ -1,9 +1,0 @@
-// @flow
-
-export default (html: string) => {
-    return html
-        .replace(/&/g, '&amp;')
-        .replace(/"/g, '&quot;')
-        .replace(/</g, '&lt;')
-        .replace(/>/g, '&gt;');
-}