+++ packages/core/mashroom/src/server/routes/mashroom/admin/info_overview_route.js
@@ -1,7 +1,7 @@
 // @flow
 
 import infoTemplate from './template';
-import jsonToHtml from './json_to_html';
+import {jsonToHtml} from '@mashroom/mashroom-utils/lib/html_utils';
 
 import type {MashroomPluginContext, ExpressRequest, ExpressResponse} from '../../../../../type-definitions';