+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [unreleased]
 
+ * Portal: Made it possible to use environment variables in the plugin definition of remote portal apps as well
  * Portal: Call the *willBeRemoved* lifecycle hook of all apps on page unload; this gives the app a chance to do some
    cleanup or persist its state properly
  * Portal: Made *defaultTheme* and *defaultLayout* in the site configuration optional and derive it from